+++ src/main/java/org/plumelib/util/AbstractMostlySingletonSet.java
@@ -7,6 +7,7 @@ import java.util.NoSuchElementException;
 import java.util.Set;
 import org.checkerframework.checker.index.qual.NonNegative;
 import org.checkerframework.checker.lock.qual.GuardSatisfied;
+import org.checkerframework.checker.nullness.qual.KeyForBottom;
 import org.checkerframework.checker.nullness.qual.Nullable;
 import org.checkerframework.checker.nullness.qual.PolyNull;
 import org.checkerframework.checker.signedness.qual.PolySigned;
@@ -38,8 +39,10 @@ public abstract class AbstractMostlySingletonSet<T extends Object> implements Se
 
   /** The current state. */
   protected State state;
+
   /** The current value, non-null when the state is SINGLETON. */
   protected @Nullable T value;
+
   /** The wrapped set, non-null when the state is ANY. */
   protected @Nullable Set<T> set;
 
@@ -170,7 +173,7 @@ public abstract class AbstractMostlySingletonSet<T extends Object> implements Se
 
   @SuppressWarnings({"nullness:unneeded.suppression", "keyfor:override.return"}) // temporary
   @Override
-  public <S> @Nullable S[] toArray(@PolyNull S[] a) {
+  public <@KeyForBottom S> @Nullable S[] toArray(@PolyNull S[] a) {
     throw new UnsupportedOperationException();
   }